requirements.txt
opencv-python-headless==4.13.0.92
pandas==2.3.3
matplotlib==3.10.8
seaborn==0.13.2
PyYAML==6.0.3
ultralytics==8.4.39
ipykernel>=6.29
pillow
requests
scipy
tqdm
psutil
